Restructure commands and code
