+++ src/commands.rs
@@ -1,28 +1,18 @@
-use add::AddArgs;
-use clap::{Parser, Subcommand};
+use clap::Subcommand;
 
 mod decks;
-use decks::*;
-mod cards;
+use decks::{DeckCommands, DeckArgs};
+
+mod add;
+use add::AddArgs;
+mod init;
 use init::InitArgs;
+mod remove;
 use remove::RemoveArgs;
+mod list;
+
+mod study;
 use study::StudyArgs;
-
-#[derive(Parser, Debug, Clone)]
-pub struct DeckArgs {
-    #[command(subcommand)]
-    cmd: DeckCommands,
-
-    deck_id: Option<String>,
-}
-#[derive(Subcommand, Debug, Clone)]
-pub enum DeckCommands {
-    #[command(name = "add", alias = "a", about = "Add a card to a deck")]
-    Add(cards::add::AddArgs),
-
-    #[command(name = "remove", alias = "rm", about = "Remove a card from a deck")]
-    Remove(cards::remove::RemoveArgs),
-}
 
 #[derive(Subcommand, Debug, Clone)]
 pub enum Commands {
@@ -49,7 +39,12 @@ pub enum Commands {
     #[command(name = "study", alias = "s", about = "Study a deck")]
     Study(StudyArgs),
 
-    #[command(name = "deck", alias = "d", about = "Manage cards in a deck")]
+    #[command(
+        name = "deck",
+        alias = "d",
+        about = "Manage cards in a deck",
+        arg_required_else_help = true
+    )]
     Deck(DeckArgs),
 }
 
@@ -59,10 +54,11 @@ pub fn run_command(cmd: Commands) {
         Commands::Init(args) => init::run(args),
         Commands::Add(args) => add::run(args),
         Commands::Remove(args) => remove::run(args),
-        Commands::Study(args) => study::run(args),
         Commands::Deck(args) => match args.cmd {
-            DeckCommands::Add(sub_args) => cards::add::run(args.deck_id, sub_args),
-            DeckCommands::Remove(sub_args) => cards::remove::run(args.deck_id, sub_args),
+            DeckCommands::Add(sub_args) => decks::add::run(Some(args.deck_id), sub_args),
+            DeckCommands::Remove(sub_args) => decks::remove::run(Some(args.deck_id), sub_args),
         },
+
+        Commands::Study(args) => study::run(args),
     }
 }
