--- src/commands/cards/mod.rs
@@ -1,2 +1,0 @@
-pub mod add;
-pub mod remove;
