+++ src/commands/decks/mod.rs
@@ -1,5 +1,24 @@
-pub mod list;
-pub mod init;
+use clap::{Parser, Subcommand};
+
 pub mod add;
 pub mod remove;
-pub mod study;
+
+#[derive(Parser, Debug, Clone)]
+pub struct DeckArgs {
+    pub deck_id: String,
+
+    #[command(subcommand)]
+    pub cmd: DeckCommands,
+}
+#[derive(Subcommand, Debug, Clone)]
+pub enum DeckCommands {
+    #[command(
+        name = "add",
+        alias = "a",
+        about = "Add an existing deck to the config"
+    )]
+    Add(add::AddArgs),
+
+    #[command(name = "remove", alias = "rm", about = "Remove a deck from the config")]
+    Remove(remove::RemoveArgs),
+}
