+++ src/commands/decks/remove.rs
@@ -1,69 +1,77 @@
-use std::path::Path;
+use std::{fs, path::Path};
 
 use clap::Parser;
-use inquire::{Confirm, Select};
+use inquire::{Confirm, Select, Text};
+
+use crate::{config::get_config, types::Deck};
 
 #[derive(Parser, Debug, Clone)]
 pub struct RemoveArgs {
-    id: Option<String>,
+    pub id: Option<String>,
 }
 
-pub fn run(args: RemoveArgs) {
-    let mut config = crate::config::get_config();
+pub fn run(_deck_id: Option<String>, args: RemoveArgs) {
+    let config = get_config();
 
-    if config.decks.is_empty() {
-        println!("No decks found.");
-        return;
+    let deck_id = match _deck_id {
+        Some(id) => {
+            if !config.decks.iter().any(|deck| deck.id == id) {
+                panic!("Deck not found: {}", id);
+            }
+
+            id
+        }
+        None => match Select::new("Select a deck", config.decks.clone()).prompt() {
+            Ok(selection) => selection.id,
+            Err(err) => {
+                panic!("Could not select deck: {}", err);
+            }
+        },
+    };
+
+    let deck_entry = config.decks.iter().find(|deck| deck.id == deck_id).unwrap();
+
+    // Check if the deck is in path
+    let deck_path = Path::new(deck_entry.path.as_str()).join("deck.json");
+    if !deck_path.exists() {
+        panic!("Deck not found in path: {}", deck_path.display());
     }
 
-    let id = match args.id {
-        Some(id) => id,
-        None => {
-            let mut ids = Vec::new();
-            for deck in &config.decks {
-                ids.push(deck.id.clone());
+    let mut new_deck = match fs::read_to_string(&deck_path) {
+        Ok(contents) => match serde_json::from_str::<Deck>(&contents) {
+            Ok(deck) => deck,
+            Err(err) => {
+                panic!("Could not parse deck file: {}", err);
             }
-
-            if let Ok(id) = Select::new("Which deck do you want to remove?", ids).prompt() {
-                id
-            } else {
-                panic!("Could not get ID for deck to remove.");
-            }
+        },
+        Err(err) => {
+            panic!("Could not read deck file: {}", err);
         }
     };
 
-    let path = config
-        .decks
-        .iter()
-        .find(|deck| deck.id == id)
-        .unwrap()
-        .path
-        .clone();
+    let id = match args.id {
+        Some(id) => id,
+        None => match Text::new("Enter a unique ID for the card").prompt() {
+            Ok(id) => id,
+            Err(err) => {
+                panic!("Could not get ID for card: {}", err);
+            }
+        },
+    };
 
-    config.decks.retain(|deck| deck.id != id);
+    new_deck.cards.retain(|card| card.id != id);
 
-    crate::config::save_config(config);
-
-    if !Path::new(&path).exists() {
-        return;
-    }
-    
-    if let Ok(remove_dir) = Confirm::new("Do you want to remove the deck directory as well?")
-        .with_default(false)
-        .prompt()
-    {
-        if !remove_dir {
-            return;
+    if match Confirm::new("Are you sure you want to remove this card?").prompt() {
+        Ok(confirmation) => confirmation,
+        Err(err) => {
+            panic!("Could not get confirmation: {}", err);
         }
-    } else {
-        panic!("Failed to get confirmation for removing the deck directory.");
-    }
-
-    if let Err(err) = std::fs::remove_dir_all(&path) {
-        if err.kind() == std::io::ErrorKind::NotFound {
-            return;
+    } {
+        match fs::write(&deck_path, serde_json::to_string(&new_deck).unwrap()) {
+            Ok(_) => {}
+            Err(err) => {
+                panic!("Could not write deck file: {}", err);
+            }
         }
-
-        panic!("Could not remove directory: {}", err);
     }
 }
